# docs/swagger.yaml
definitions:
...
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      msg:
        type: string
    type: object
...
  service.Order:
    properties:
      id:
        type: integer
      product_id:
        type: integer
      state:
        type: integer
      user_id:
        type: string
    type: object
...
  service.User:
    properties:
      id:
        type: string
      nick_name:
        type: string
      user_name:
        type: string
    type: object
info:
  contact: {}
  description: 超级工具合集
  license: {}
  title: superTools
  version: "1.0"
paths:
...
  /api/v1/mall/orders:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.Order'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取订单列表
      tags:
      - mall
    post:
      parameters:
      - description: 订单id
        in: body
        name: id
        required: true
        schema:
          type: integer
      - description: 用户id
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: 商品id
        in: body
        name: product_id
        required: true
        schema:
          type: integer
      - description: 订单状态
        in: body
        name: state
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: "1"
          schema:
            $ref: '#/definitions/int'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 新增订单
      tags:
      - mall
    put:
      parameters:
      - description: 订单id
        in: body
        name: id
        required: true
        schema:
          type: integer
      - description: 用户id
        in: body
        name: user_id
        required: true
        schema:
          type: integer
      - description: 商品id
        in: body
        name: product_id
        required: true
        schema:
          type: integer
      - description: 订单状态
        in: body
        name: state
        required: true
        schema:
          type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/string'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 更新订单
      tags:
      - mall
...
  /api/v1/user/signin:
    post:
      parameters:
      - description: 用户名
        in: body
        name: user_name
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.User'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 用户登录
      tags:
      - user
...
swagger: "2.0"